+++ test/tools/chainReducers.js
@@ -1,0 +1,9 @@
+import { expect } from 'chai';
+import { chainReducers } from '../../src/index';
+
+describe('chainReducers', () => {
+  it('should work', async function () {
+
+    return;
+  });
+});