+++ docs/changelog.md
@@ -10,3 +10,8 @@
 
 ### 2.0.2
 - Fixed failing tests
+
+### 3.0.0
+- Refactored tests
+- Setting default value in "connect" function
+- "connect" function throws errors when path has not been found